# Implement cong in Julia