@vchuravy 2023-06-19
I think the challenge is that we want randomness that is independent of the task state, since we don't have "scheduler state" outside of the ptls that's the only option.

I am more annoyed how slow this is due to the mod operation.