@oscardssmith 2023-06-19
Can we guarantee that the size of the heaps are powers of 2? If so, you can just use an `&`.