@vchuravy 2023-06-19
It's the number of heaps and not the size of them, and that is equal to the number of threads. Which often is a power of two, but not guaranteed