@oscardssmith 2023-06-19
What about using `Base.MultiplicativeInverses` to precompute the multiplicativeinverse of the number of threads?