@IanButterworth 2024-07-15
Is this a valid benchmark?

## this PR
```
% ./julia -t6 -q
julia> function fib(n::Int)
                  if n < 2
                      return n
                  end
                  t = Threads.@spawn fib(n - 2)
                  return fib(n - 1) + fetch(t)
              end
fib (generic function with 1 method)

julia> using BenchmarkTools

julia> @btime fib(30)
  2.098 s (6929222 allocations: 1.13 GiB)
832040

julia> @btime fib(30)
  2.029 s (6896970 allocations: 1.05 GiB)
832040
```

## master b88f64f16c4 (via juliaup)
```
% julia +nightly -t6
...
julia> @btime fib(30)
  748.326 ms (6772487 allocations: 734.01 MiB)
832040

julia> @btime fib(30)
  700.210 ms (6743099 allocations: 659.97 MiB)
832040
```