@vchuravy 2024-07-15
Probably. I haven't checked the type-stability of this code after the recent rewrite. Since the allocation increased I assume we have one.

@IanButterworth do you want to take this over?