@IanButterworth 2024-07-15
Doesn't appear to be an allocation within `cong`. And it's faster.
This PR [b5d065a](https://github.com/JuliaLang/julia/pull/50203/commits/b5d065a98968991bffa73cebb27a128604bf5f0a)
```
julia> @btime Base.Partr.cong(UInt32(100))
  2.333 ns (0 allocations: 0 bytes)
```
master
```
julia> @btime Base.Partr.cong(UInt32(100))
  5.333 ns (0 allocations: 0 bytes)
0x00000036
```

However I just locally updated this branch from master and it regressed to
```
julia> @btime Base.Partr.cong(UInt32(100))
  6.958 ns (0 allocations: 0 bytes)
0x00000061
```