@gbaraldi 2024-07-24
So looking into this together with Ian we discovered that our rng is just very bad. I also discovered that `get_ptls` is not getting optimized as it should because the declaration is behind an ifdef so the ccall transform doesn't do it.
Finally we also discovered that `rand(1:range)` performs quite on par with the current cong and better than the corrected one.
The corrected one being
```julia
function _cong(max::UInt64, seed::UInt64)
    if max < 2
        return UInt64(0), seed
    end
    mask = typemax(UInt64)
    zeros = leading_zeros(max)
    mask >>= zeros
    bits = 8 * sizeof(UInt64) - zeros
    while true
        value = UInt64(69069) * seed + UInt64(362437)
        seed = value
        x = seed & mask
        if (x < max)
            return x, seed
        end
        bits_left = zeros
        while bits_left >= bits
            value >>= bits
            x = value & mask
            if x < max
                return x, seed
            end
            bits_left -= bits
        end
    end
    return x, seed
end
```