@IanButterworth 2024-07-24
To illustrate the periodicity of `cong` on master.
```
julia> for i in 2:20
           @info i
           x = [Base.Partr.cong(UInt32(i)) for _ in 1:100];
           display(lineplot(x, width=100, height = 8))
end
```
| 2-6 | 7-11  | 12-16 |
| --- | --- | --- |
| ![Screenshot 2024-07-23 at 10 26 47 PM](https://github.com/user-attachments/assets/d1f02b80-d8d4-4ca0-9edc-dd3170ea6da5) | ![Screenshot 2024-07-23 at 10 26 59 PM](https://github.com/user-attachments/assets/b38b7984-3088-453b-be1b-947cc1f50671) | ![Screenshot 2024-07-23 at 10 27 15 PM](https://github.com/user-attachments/assets/9b8bab7d-0d74-497b-9425-f9fc993c2696) |